pinecone_storage/requirements.txt
Flask==2.3.2
pinecone-client[grpc]
